avtovin_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
httpx==0.27.0
firebase-admin==6.6.0
python-multipart==0.0.12
websockets==13.1
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
